# apps/web/lib/api/openapi.json
{"openapi":"3.0.1","info":{"title":"IHRMS API","description":"IHRMS backend. Contract: docs/api-contract.md","version":"0.1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearer-jwt":[]}],"paths":{"/teams/{id}/manager":{"put":{"tags":["teams-controller"],"operationId":"assignManager","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignMemberRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssignMemberResult"}}}}}}},"/teams/{id}/hr":{"put":{"tags":["teams-controller"],"operationId":"assignHr","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignMemberRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AssignMemberResult"}}}}}}},"/storage/blobs/{token}":{"get":{"tags":["storage-controller"],"operationId":"download","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}},"put":{"tags":["storage-controller"],"operationId":"upload","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"byte"}}}},"responses":{"200":{"description":"OK"}}}},"/me/onboarding/sections/{key}":{"put":{"tags":["onboarding-controller"],"operationId":"saveSection","parameters":[{"name":"key","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveSectionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileSectionView"}}}}}}},"/teams":{"get":{"tags":["teams-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TeamSummaryView"}}}}}}},"post":{"tags":["teams-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTeamRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamDetailView"}}}}}}},"/me/onboarding/submit":{"post":{"tags":["onboarding-controller"],"operationId":"submit","responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OnboardingDashboard"}}}}}}},"/me/onboarding/documents":{"post":{"tags":["onboarding-controller"],"operationId":"requestUpload","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentUploadRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PresignedUpload"}}}}}}},"/me/onboarding/documents/{id}/confirm":{"post":{"tags":["onboarding-controller"],"operationId":"confirm","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentView"}}}}}}},"/manager/notifications/{id}/read":{"post":{"tags":["manager-controller"],"operationId":"markRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationView"}}}}}}},"/manager/notifications/read-all":{"post":{"tags":["manager-controller"],"operationId":"markAllRead","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationFeed"}}}}}}},"/manager/approvals/{id}/reject":{"post":{"tags":["manager-controller"],"operationId":"reject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectApprovalRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApprovalView"}}}}}}},"/manager/approvals/{id}/approve":{"post":{"tags":["manager-controller"],"operationId":"approve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApprovalView"}}}}}}},"/employees":{"get":{"tags":["employees-controller"],"operationId":"list_1","parameters":[{"name":"search","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["INVITED","IN_PROGRESS","SUBMITTED","HR_VERIFIED","APPROVED","REJECTED"]}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeePage"}}}}}},"post":{"tags":["employees-controller"],"operationId":"onboard","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardEmployeeRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OnboardEmployeeResult"}}}}}}},"/employees/{id}/route-to-manager":{"post":{"tags":["review-controller"],"operationId":"routeToManager","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteToManagerRequest"}}}},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RouteToManagerResult"}}}}}}},"/companies":{"get":{"tags":["companies-controller"],"operationId":"list_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CompanySummaryView"}}}}}}},"post":{"tags":["companies-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanyDetailView"}}}}}}},"/companies/{id}/admin":{"post":{"tags":["companies-controller"],"operationId":"provisionAdmin","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProvisionAdminRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProvisionAdminResult"}}}}}}},"/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","parameters":[{"name":"ihrms_refresh","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResult"}}}}}}},"/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OkResponse"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StaffLoginRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResult"}}}}}}},"/auth/employee/verify-otp":{"post":{"tags":["auth-controller"],"operationId":"verifyOtp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOtpVerifyRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResult"}}}}}}},"/auth/employee/request-otp":{"post":{"tags":["auth-controller"],"operationId":"requestOtp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOtpRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OtpRequestResult"}}}}}}},"/teams/{id}":{"get":{"tags":["teams-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamDetailView"}}}}}},"delete":{"tags":["teams-controller"],"operationId":"remove","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"boolean"}}}}}}},"patch":{"tags":["teams-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTeamRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TeamDetailView"}}}}}}},"/employees/{id}/sections/{key}":{"patch":{"tags":["review-controller"],"operationId":"reviewSection","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"key","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeRecordView"}}}}}}},"/employees/{id}/documents/{documentId}":{"patch":{"tags":["review-controller"],"operationId":"reviewDocument","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"documentId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeRecordView"}}}}}}},"/companies/{id}":{"get":{"tags":["companies-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanyDetailView"}}}}}},"patch":{"tags":["companies-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanyDetailView"}}}}}}},"/teams/assignable-users":{"get":{"tags":["teams-controller"],"operationId":"assignable","parameters":[{"name":"role","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TeamMemberView"}}}}}}}},"/me/onboarding":{"get":{"tags":["onboarding-controller"],"operationId":"dashboard","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OnboardingDashboard"}}}}}}},"/me/onboarding/documents/{id}/url":{"get":{"tags":["onboarding-controller"],"operationId":"viewUrl","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PresignedView"}}}}}}},"/manager/notifications":{"get":{"tags":["manager-controller"],"operationId":"notifications","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationFeed"}}}}}}},"/manager/approvals":{"get":{"tags":["manager-controller"],"operationId":"approvals","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApprovalView"}}}}}}}},"/manager/approvals/{id}/record":{"get":{"tags":["manager-controller"],"operationId":"approvalRecord","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeRecordView"}}}}}}},"/manager/approvals/history":{"get":{"tags":["manager-controller"],"operationId":"approvalHistory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApprovalView"}}}}}}}},"/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/employees/{id}/record":{"get":{"tags":["review-controller"],"operationId":"record","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeRecordView"}}}}}}},"/employees/lookup/{employeeCode}":{"get":{"tags":["review-controller"],"operationId":"lookup","parameters":[{"name":"employeeCode","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeRecordView"}}}}}}},"/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionView"}}}}}}},"/audit":{"get":{"tags":["audit-controller"],"operationId":"query","parameters":[{"name":"companyId","in":"query","required":false,"schema":{"type":"string"}},{"name":"action","in":"query","required":false,"schema":{"type":"string"}},{"name":"actorType","in":"query","required":false,"schema":{"type":"string"}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuditPage"}}}}}}},"/me/onboarding/documents/{id}":{"delete":{"tags":["onboarding-controller"],"operationId":"deleteDocument","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OnboardingDashboard"}}}}}}}},"components":{"schemas":{"AssignMemberRequest":{"type":"object","properties":{"userId":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"}}},"AssignMemberResult":{"type":"object","properties":{"team":{"$ref":"#/components/schemas/TeamDetailView"},"devPassword":{"type":"string"}}},"TeamDetailView":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"hr":{"$ref":"#/components/schemas/TeamMemberView"},"manager":{"$ref":"#/components/schemas/TeamMemberView"},"memberCount":{"type":"integer","format":"int64"},"createdAt":{"type":"string"},"members":{"type":"array","items":{"$ref":"#/components/schemas/TeamMemberView"}}}},"TeamMemberView":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["SUPER_ADMIN","COMPANY_ADMIN","HR","MANAGER"]},"status":{"type":"string"}}},"SaveSectionRequest":{"required":["data"],"type":"object","properties":{"data":{"type":"object","additionalProperties":{"type":"object"}}}},"ProfileSectionView":{"type":"object","properties":{"key":{"type":"string","enum":["PERSONAL","BACKGROUND","GOVERNMENT"]},"data":{"type":"object","additionalProperties":{"type":"object"}},"status":{"type":"string","enum":["DRAFT","SUBMITTED","VERIFIED","REJECTED"]},"updatedAt":{"type":"string"}}},"CreateTeamRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":120,"minLength":2,"type":"string"}}},"DocumentView":{"type":"object","properties":{"id":{"type":"string"},"sectionKey":{"type":"string","enum":["PERSONAL","BACKGROUND","GOVERNMENT"]},"docType":{"type":"string","enum":["EXPERIENCE_LETTER","PAN","AADHAAR","BGV_DOCUMENT","OTHER"]},"fileName":{"type":"string"},"mimeType":{"type":"string"},"sha256":{"type":"string"},"status":{"type":"string","enum":["PENDING","UPLOADED","VERIFIED","REJECTED"]},"uploadedAt":{"type":"string"}}},"OnboardingDashboard":{"type":"object","properties":{"employeeCode":{"type":"string"},"email":{"type":"string"},"status":{"type":"string","enum":["INVITED","IN_PROGRESS","SUBMITTED","HR_VERIFIED","APPROVED","REJECTED"]},"sections":{"type":"array","items":{"$ref":"#/components/schemas/ProfileSectionView"}},"documents":{"type":"array","items":{"$ref":"#/components/schemas/DocumentView"}}}},"DocumentUploadRequest":{"required":["docType","fileName","mimeType","sectionKey"],"type":"object","properties":{"sectionKey":{"type":"string","enum":["PERSONAL","BACKGROUND","GOVERNMENT"]},"docType":{"type":"string","enum":["EXPERIENCE_LETTER","PAN","AADHAAR","BGV_DOCUMENT","OTHER"]},"fileName":{"maxLength":255,"minLength":0,"type":"string"},"mimeType":{"pattern":"application/pdf|image/png|image/jpeg","type":"string"},"sizeBytes":{"maximum":10485760,"type":"integer","format":"int64"}}},"PresignedUpload":{"type":"object","properties":{"documentId":{"type":"string"},"uploadUrl":{"type":"string"},"method":{"type":"string"},"headers":{"type":"object","additionalProperties":{"type":"string"}},"expiresInSeconds":{"type":"integer","format":"int32"}}},"NotificationView":{"type":"object","properties":{"id":{"type":"string"},"type":{"type":"string","enum":["EMPLOYEE_ONBOARDED","EMPLOYEE_SUBMITTED","APPROVAL_REQUESTED","EMPLOYEE_APPROVED","EMPLOYEE_REJECTED"]},"employeeId":{"type":"string"},"employeeCode":{"type":"string"},"read":{"type":"boolean"},"createdAt":{"type":"string"}}},"NotificationFeed":{"type":"object","properties":{"notifications":{"type":"array","items":{"$ref":"#/components/schemas/NotificationView"}},"unreadCount":{"type":"integer","format":"int64"}}},"RejectApprovalRequest":{"required":["note"],"type":"object","properties":{"note":{"maxLength":500,"minLength":0,"type":"string"}}},"ApprovalView":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED"]},"note":{"type":"string"},"submittedAt":{"type":"string"},"decidedAt":{"type":"string"},"employeeCode":{"type":"string"},"employeeEmail":{"type":"string"},"employeeStatus":{"type":"string","enum":["INVITED","IN_PROGRESS","SUBMITTED","HR_VERIFIED","APPROVED","REJECTED"]},"hrName":{"type":"string"}}},"OnboardEmployeeRequest":{"required":["dateOfJoining","designation","email","fullName"],"type":"object","properties":{"fullName":{"maxLength":120,"minLength":0,"type":"string"},"email":{"type":"string"},"designation":{"maxLength":120,"minLength":0,"type":"string"},"dateOfJoining":{"type":"string","format":"date"}}},"EmployeeSummaryView":{"type":"object","properties":{"id":{"type":"string"},"employeeCode":{"type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"designation":{"type":"string"},"dateOfJoining":{"type":"string"},"status":{"type":"string","enum":["INVITED","IN_PROGRESS","SUBMITTED","HR_VERIFIED","APPROVED","REJECTED"]},"createdAt":{"type":"string"}}},"OnboardEmployeeResult":{"type":"object","properties":{"employee":{"$ref":"#/components/schemas/EmployeeSummaryView"},"loginUrl":{"type":"string"}}},"RouteToManagerRequest":{"type":"object","properties":{"note":{"maxLength":500,"minLength":0,"type":"string"}}},"RouteToManagerResult":{"type":"object","properties":{"employeeCode":{"type":"string"},"status":{"type":"string","enum":["INVITED","IN_PROGRESS","SUBMITTED","HR_VERIFIED","APPROVED","REJECTED"]},"approvalRequestId":{"type":"string"},"managerName":{"type":"string"}}},"CreateCompanyRequest":{"required":["code","name"],"type":"object","properties":{"name":{"maxLength":120,"minLength":2,"type":"string"},"code":{"type":"string"}}},"CompanyAdminView":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string"}}},"CompanyDetailView":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"code":{"type":"string"},"status":{"type":"string"},"teamCount":{"type":"integer","format":"int64"},"employeeCount":{"type":"integer","format":"int64"},"hasAdmin":{"type":"boolean"},"createdAt":{"type":"string"},"admin":{"$ref":"#/components/schemas/CompanyAdminView"}}},"ProvisionAdminRequest":{"required":["email","name"],"type":"object","properties":{"name":{"maxLength":120,"minLength":2,"type":"string"},"email":{"type":"string"}}},"ProvisionAdminResult":{"type":"object","properties":{"admin":{"$ref":"#/components/schemas/CompanyAdminView"},"devPassword":{"type":"string"}}},"AuthResult":{"type":"object","properties":{"accessToken":{"type":"string"},"session":{"$ref":"#/components/schemas/SessionView"}}},"SessionView":{"type":"object"},"OkResponse":{"type":"object","properties":{"ok":{"type":"boolean"}}},"StaffLoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"EmployeeOtpVerifyRequest":{"required":["email","otp"],"type":"object","properties":{"email":{"type":"string"},"otp":{"pattern":"\\d{6}","type":"string"}}},"EmployeeOtpRequest":{"required":["email","fullName"],"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string"}}},"OtpRequestResult":{"type":"object","properties":{"sent":{"type":"boolean"},"expiresInSeconds":{"type":"integer","format":"int32"},"devOtp":{"type":"string"}}},"UpdateTeamRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":120,"minLength":2,"type":"string"}}},"ReviewRequest":{"required":["decision"],"type":"object","properties":{"decision":{"pattern":"VERIFIED|REJECTED","type":"string","enum":["VERIFIED","REJECTED"]},"reason":{"maxLength":500,"minLength":0,"type":"string"}}},"EmployeeRecordView":{"type":"object","properties":{"id":{"type":"string"},"employeeCode":{"type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"designation":{"type":"string"},"dateOfJoining":{"type":"string"},"status":{"type":"string","enum":["INVITED","IN_PROGRESS","SUBMITTED","HR_VERIFIED","APPROVED","REJECTED"]},"reviewComplete":{"type":"boolean"},"sections":{"type":"array","items":{"$ref":"#/components/schemas/RecordSection"}},"documents":{"type":"array","items":{"$ref":"#/components/schemas/RecordDocument"}}}},"RecordDocument":{"type":"object","properties":{"id":{"type":"string"},"sectionKey":{"type":"string","enum":["PERSONAL","BACKGROUND","GOVERNMENT"]},"docType":{"type":"string","enum":["EXPERIENCE_LETTER","PAN","AADHAAR","BGV_DOCUMENT","OTHER"]},"fileName":{"type":"string"},"mimeType":{"type":"string"},"sha256":{"type":"string"},"status":{"type":"string","enum":["PENDING","UPLOADED","VERIFIED","REJECTED"]},"uploadedAt":{"type":"string"},"viewUrl":{"type":"string"}}},"RecordSection":{"type":"object","properties":{"key":{"type":"string","enum":["PERSONAL","BACKGROUND","GOVERNMENT"]},"data":{"type":"object","additionalProperties":{"type":"object"}},"status":{"type":"string","enum":["DRAFT","SUBMITTED","VERIFIED","REJECTED"]},"updatedAt":{"type":"string"}}},"UpdateCompanyRequest":{"type":"object","properties":{"name":{"maxLength":120,"minLength":2,"type":"string"},"status":{"pattern":"ACTIVE|SUSPENDED","type":"string"}}},"TeamSummaryView":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"hr":{"$ref":"#/components/schemas/TeamMemberView"},"manager":{"$ref":"#/components/schemas/TeamMemberView"},"memberCount":{"type":"integer","format":"int64"},"createdAt":{"type":"string"}}},"PresignedView":{"type":"object","properties":{"url":{"type":"string"},"expiresInSeconds":{"type":"integer","format":"int32"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"EmployeePage":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeSummaryView"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"CompanySummaryView":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"code":{"type":"string"},"status":{"type":"string"},"teamCount":{"type":"integer","format":"int64"},"employeeCount":{"type":"integer","format":"int64"},"hasAdmin":{"type":"boolean"},"createdAt":{"type":"string"}}},"AuditLogView":{"type":"object","properties":{"id":{"type":"string"},"companyId":{"type":"string"},"actorType":{"type":"string"},"actorId":{"type":"string"},"actorLabel":{"type":"string"},"action":{"type":"string"},"targetType":{"type":"string"},"targetId":{"type":"string"},"metadata":{"type":"object","additionalProperties":{"type":"object"}},"ipAddress":{"type":"string"},"createdAt":{"type":"string"}}},"AuditPage":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/AuditLogView"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}}},"securitySchemes":{"bearer-jwt":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}